Use std::function for RecentMessagesApi
